+++ dotfiles/.config/quickshell/shell.qml
@@ -3,10 +3,12 @@ import "WelcomeApp"
 import "PowerApp"
 import "SidebarApp"
 import "CalendarApp"
+import "WallpaperApp"
 
 ShellRoot {
     WelcomeWindow {}
     PowerWindow {}
     SidebarWindow {}
     CalendarWindow {}
+    WallpaperWindow {}
 }