+++ setup/preflight-arch.sh
@@ -102,6 +102,6 @@ _checkAURHelper
 # Uninstall swww if exists. To be replaced with awww in the next steps
 # --------------------------------------------------------------
 
-if command -v swww &> /dev/null || pacman -Qq swww &> /dev/null; then
-    sudo pacman -Rns --noconfirm swww
-fi
+sudo pacman -Rns --noconfirm swww
+yay --noconfirm -Rns waypaper-git
+yay --noconfirm -Rns waypaper